Add basic file upload functionality
Added basic ability to choose files from a local file browser and store these as csv files in the webapp/resources folder in the project files. This is temporary; the data is not retrieved correctly and the folder location is likely to change but the files are created with the correct names.