Created reference page.
The reference page is the page users are taken to after selecting a run from the dashboard, originally I was using the manage run page as the target but the reference page has now been obtained and I've been able to implement it. this page contains a number of buttons that the user can use to display weights, export data, and perform a number of other IPS related processes.