Fixed error with app methods
Error in the app methods with the directory of the data being referenced. Need to speak to adnan about how to do this as he recommended file paths be treated as if coming from root. This change has been made to make the reference from the .py file as this results in the functionality working but test failing