Add error message display functionality
This commit adds a standard DSS error box to the top of the page if any
errors are detected when trying to perform an upload. Additionally, any
file that has caused an error will now display a message underneath
in red colour indicating whether the field was not filled in or if the
file chosen was not a csv file.
The form now also will not progress if a file other than a csv is
uploaded, whereas before the file browser would simply default to search
for csv's and not validate them further.