Add session tracking on the new_run_2 page and write weights flask app.
In order to persist data on new_run_2 session info is collected and used. This allows us to maintain the user's input when entered incorrectly in order to give guidance at rectifying the error and to reduce the need to input the same data multiple times.

The weights decorator has been modified and written in a more flask oriented way enabling more controll over the application as opposed to using html redirects from the web page.