Add unit tests
Added 3 unit tests for the new_run_3 web page. The tests set up a
testing app client and sends requests to this app, then test for expected
results. In total, this tests for expected text in the standard URL with
a GET request, lack of error text in a GET request and the correct error
text when no files are uploaded in a POST request.