Edit tests and remove call to flash_errors
Edited one pytest so that both types of error are validated as not being
present on the basic page. Removed reference to flash_errors which will
eventually be deleted. Changed validation in the main app call to use
validate_on_submit. Still need to get request type so as not to display
the error box on the basic page.