Add method for editing run steps
New method added to enable editing of run steps (this will enable the UI to update the manage run screen when a step status needs changing). Functionality alos added for creating the initial run steps which has been coupled in with the creation of a new run (New run must have 14 Steps associated with it).