Tests written for manager run (reference).html) page.
A series of flask tests were written to ensure correct re-directions were in place when interacting with the manage run page (reference.html). The test include both GET and POST methods acting on the page and accommodate for all currently implemented functionality. (The starting of the run itself is not implemented.)