Implement creation of new pv sets from existing set templates
When creating a new run through the UI, new_run_4 and 5 will now create
a new set of process variables that has the same pv's as a template
chosen by a radio button. When moving to new_run_5, the chosen template
will be used to create a new set which has the run_id of the current
session. The new set will have the same pv values as the template set
and is stored in the API gateway placeholder database.