Implement partial modal (popup box) for editing PV's
Implemented a button that creates a modal, which is a box overlayed
on top of the current page. This box currently has a text input box
with no functionality except text entry and a placeholder value. This
will be used to retrieve edits to process variables on page new_run_5.