Update to current progress for visibility
Committing this so that latest state is available while I'm in annual
leave. We have a html/Jinja2 table in a form that takes user input and
edits the variables in the table. This form currently doesn't pass the
data in the table back to Flask. When we can achieve this, the process
variables functionality will be done aside from jsonifying that data and
sending that to the appropriate API backend function.