Use json style js passing to implement pv sets functionality
Creating new pv sets now works correctly; a json style object is passed
and the new set is created with the data included. Receiving this data
back from the database to be displayed on new_run_5 not currently
working; Reason and name are received but changes to the pv_content are
not taken from the template. Instead, all examples show 'Description'.