Fix error in how new_run_5 detected and used pv_content
New_run_5 now also contains a hidden column on the main page containing
the process variables content. This column now correctly receives data
from the table records passed into it and so are passed to the modal
correctly. Previously, the column was not hidden and contained the
explicit string 'Description', so no data was being received by that
column from the chosen pv_set template.