Add tweak to enable edit run to work as intended
When creating process variables, the code now checks whether
the run_id already exists. If not, a set pv_set is created
and is filled appropriately with the edited pv's. If the
run_id does exist, the previous pv's associated with that
id are deleted and replaced with the edited data from the
html table.