Update html button references, remove faulty test
Buttons on various pages have been changed so that their references work correctly, such as giving the correct reference for system_info on all the new_run pages that are used. The buttons on the export and manage run page were also tested and edited, as well as tweaking the naming on the Export page to read 'Manage Run' and 'Export' instead of 'Runs' and 'Exports'.