Partly implement Flask logging, remove old html files
Logging has been partially implemented; functionality exists and is
currently being added to and fleshed out but logging has not yet been
added to each line of code that will need it. Also currently logs only
to a txt file in the system. Old HTML templates new_run_6 to 9 + end
removed as they are not used.