Implement generic logging template, created in __init__.py
Flask logger is now created with configured formatting in the init file.
The formatter ensures that every logged message will contain the time
the message was created, the name of the function (endpoint) that logged
the message, the severity level and the message itself. Uses a
RotatingFileHandler which creates up to 2 new logs when the previous log
reaches 2000 bytes in size (this is an arbitrary size).
This task will include full implementation of logging messages after
the requirements are known.