Implement initial RabbitMq logging system
This commit adds a custom logging StreamHandler setup to use the logger
to emit messages to RabbitMq. On the VDI, this has shown to push
messages from the dashboard logging system onto a desired Rabbit queue.
Currently, the system doesn't work on more than 1 page; dashboard logs
are sent successfully but will freeze when submitting a new request.
This happens both when moving to a new page and when resubmitting a
request to the same page.