Update requirements.txt
This commit adds setuptools, pika and requests to the
requirements.txt file. Now, assuming that Python and pip
are installed and are added to the system path, it is
possible to install everything required by the UI at the
same time in PyCharm by calling pip install -r requirements.txt.
Once this is done, the environment still needs to be set up
correctly but all dependencies will be installed.