Milestone - Process variables work correctly through VDI UI
All process variables edits required have now been made. Testing has
allowed Tom and I to create a new run all the way through to creation
of a new set of fully populated process variables, which are all
pushed to and retrieved from the SQL server database. Process variable
ID is now accounted for, so process variables are sorted when displayed
on new_run_5 and in the database.

The data input (new_run_3) still needs to be completed. Currently not
storing the input data anywhere; this is our next task.

This commit also includes smaller edits, such as changing the number
of steps shown in the new_run HTML pages from 7 to the correct value, 5.