Complete export data functionality
This commit includes complete export data functionality in the UI, which
retrieves data from SQL and produces a zipped csv. A run can now be
created and the run_id generated can be used to retrieve appropriate
data from SQL with that id. However, this is currently manually added
to SQL as we are unsure how the import is supposed to happen.

Other changes include limiting the entry box for a run_name to 30
characters (preventing issues we had previously with long names) and
disabling logging temporarily (since RabbitMQ has been turned off).