Restore error reports and fix run selected
Run selected button on manage_run page now works correctly to receive
information from SQL about the status of each step. Steps with errors
or warnings now create a clickable button which displays the messages
in a modal.

Clicking Run Selected will now produce a timed sub-process which
periodically updates the SQL table with more information. Next step
is to automatically update the inbformation displayed on the page
as it changes in SQL.